refactor: point in-repo callers at the moved share/rename/mapper endpoints
Update the desk JS callers (share sidebar, form toolbar rename dialog,
create_new mapped docs) and test imports to the canonical
frappe.core.api.document paths and names.
