+++ frappe/tests/test_rename_doc.py
@@ -8,10 +8,11 @@ from random import choice, sample
 from unittest.mock import patch
 
 import frappe
+from frappe.core.api.document import update_document_title
 from frappe.core.doctype.doctype.test_doctype import new_doctype
 from frappe.exceptions import DoesNotExistError
 from frappe.model.base_document import get_controller
-from frappe.model.rename_doc import bulk_rename, update_document_title
+from frappe.model.rename_doc import bulk_rename
 from frappe.modules.utils import get_doc_path
 from frappe.tests import IntegrationTestCase
 from frappe.utils import add_to_date, now
